Move a part of code to component Result
